KnowMap_Project milestone_1/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
bcrypt==4.0.1  # Direct bcrypt
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.0